Sprint 23 Исправления
