+++ app/src/main/java/com/example/playlistmaker/sharing/domain/api/ExternalNavigator.kt
@@ -4,6 +4,7 @@ import com.example.playlistmaker.sharing.domain.model.EmailData
 
 interface ExternalNavigator {
     fun shareLink(link: String?)
+    fun sharePlaylist(message: String?)
     fun openLink(agreement: String?)
     fun openEmail(emailData: EmailData)
 }
