+++ app/src/main/res/values-night/styles.xml
@@ -144,4 +144,21 @@
         <item name="cornerRadius">@dimen/dm_54dp</item>
         <item name="android:textAllCaps">false</item>
     </style>
+
+    <style name="PlaylistBottomSheetShareStyle">
+        <item name="android:textColor">@color/white</item>
+        <item name="android:textSize">@dimen/settings_content_text_size_16sp</item>
+        <item name="android:fontFamily">@font/ys_display_regular</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:layout_width">wrap_content</item>
+        <item name="android:layout_height">wrap_content</item>
+        <item name="android:gravity">center_vertical</item>
+    </style>
+
+    <style name="PlaylistBottomSheetTitleStyle">
+        <item name="android:textColor">@color/white</item>
+        <item name="android:textSize">@dimen/dm_19sp</item>
+        <item name="android:fontFamily">@font/ys_display_medium</item>
+    </style>
+
 </resources>
